Cipa/requirements.txt
Django==1.10.2
django-bootstrap3==7.1.0
django-easy-pdf==0.1.0
django-extensions==1.7.4
django-wkhtmltopdf==3.1.0
html5lib==1.0b8
MySQL-python==1.2.5
Pillow==3.4.1
pydot==1.2.3
pyparsing==2.1.10
PyPDF2==1.26.0
reportlab==3.3.0
six==1.10.0
sorl-thumbnail==12.3
webencodings==0.5
xhtml2pdf==0.0.6
